requirements.txt
opencv_python_headless
streamlit
numpy
matplotlib
Pillow
